@@ -11,35 +11,34 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^8.2.14",
-    "@angular/cdk": "^8.2.3",
+    "@angular/animations": "^9.0.2",
+    "@angular/cdk": "^9.0.1",
     "@angular/common": "^9.0.2",
-    "@angular/compiler": "^8.2.14",
-    "@angular/core": "^8.2.14",
-    "@angular/forms": "^8.2.14",
-    "@angular/material": "^8.2.3",
-    "@angular/platform-browser": "^8.2.14",
-    "@angular/platform-browser-dynamic": "^8.2.14",
-    "@angular/router": "^8.2.14",
+    "@angular/compiler": "^9.0.2",
+    "@angular/core": "^9.0.2",
+    "@angular/forms": "^9.0.2",
+    "@angular/material": "^9.0.1",
+    "@angular/platform-browser": "^9.0.2",
+    "@angular/platform-browser-dynamic": "^9.0.2",
+    "@angular/router": "^9.0.2",
     "code-prettify": "^0.1.0",
     "core-js": "^3.1.3",
     "rxjs": "6.5.4",
-    "zone.js": "~0.9.1"
+    "tslib": "^1.10.0",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular/compiler-cli": "^8.2.14",
-    "@angular-devkit/build-ng-packagr": "~0.803.25",
-    "@angular-devkit/build-angular": "~0.803.25",
-    "ng-packagr": "^5.7.1",
-    "tsickle": "0.35.0",
-    "tslib": "^1.11.0",
-    "typescript": "~3.5.3",
-    "@angular/cli": "8.3.25",
-    "@angular/language-service": "^8.2.14",
+    "@angular/compiler-cli": "^9.0.2",
+    "@angular-devkit/build-ng-packagr": "~0.900.3",
+    "@angular-devkit/build-angular": "~0.900.3",
+    "ng-packagr": "^9.0.0",
+    "typescript": "~3.7.5",
+    "@angular/cli": "9.0.3",
+    "@angular/language-service": "^9.0.2",
     "@types/jasmine": "~3.3.13",
     "@types/jasminewd2": "~2.0.6",
-    "@types/node": "~12.0.4",
-    "codelyzer": "^5.0.1",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
     "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~4.1.0",
